refactor: 拆分 AIAnalysisPanel 组件优化代码结构
主要改进：
1. 组件拆分：
   - 创建 TrendAnalysis.tsx 独立趋势分析组件（154行）
   - 创建 OptimizationAdvice.tsx 独立优化建议组件（131行）
   - 提取 utils.ts 工具函数（310行）
   - 主组件从 808 行精简到 363 行（减少 55%）

2. 代码结构优化：
   - 将数据处理逻辑从组件中分离到 utils.ts
   - 提供可复用的数据处理函数
   - 使用明确的类型定义（TrendAnalysisData, PersonalizedAdviceData）

3. 组件职责单一：
   - TrendAnalysis：专注趋势数据展示
   - OptimizationAdvice：专注建议展示
   - 主组件：专注状态管理和数据协调

4. 提升可维护性：
   - 子组件独立测试更容易
   - 代码复用性更强
   - 逻辑更清晰，修改影响范围更小

性能优势：
- 子组件可独立优化渲染
- 更小的组件粒度便于 React 优化
- 代码分割更灵活

文件结构：
components/AIAnalysisPanel/
├── index.tsx (原 AIAnalysisPanel.tsx)
├── TrendAnalysis.tsx (新增)
├── OptimizationAdvice.tsx (新增)
├── utils.ts (新增)
└── constants.ts (已存在)
